+++ magicgui/bases.py
